Create action to build binaries (#43)
This adds a Github Action to build the conformance runner binaries for
Linux and MacOS for x86_64 architectures. Building for ARM64 is planned
for a follow-up as it is not very easy to build these using Github
runners. Github does not provide native ARM64 images for Linux or Mac.

The trigger for the action is a tag being pushed to the repo. The plan
is for the tags in this repo to mirror the tags from Protobuf. Ideally,
it would be nice to poll the [Protobuf
repo](https://github.com/protocolbuffers/protobuf) for releases, but
that isn't an easy task either and could be done in a follow-up also.

Pushing an invalid tag not found on the Protobuf repo will cause the
action to fail.

Example of successful run (note: release and tags were deleted after the
run):
https://github.com/bufbuild/protobuf-conformance/actions/runs/4692167405