Use the released binaries to run conformance tests (#139)
This uses the binaries published by this repo to run the conformance
tests rather than build them via Bazel. The binaries are already being
built via Bazel as part of the release process, so there's no need to
duplicate the process. This also greatly reduces the time for CI to run.

Also removes `testci` make target as this is no longer used.