Group dependabot updates per implementation (#160)
This adds a `groups` definition for each implementation for dependabot
updates. Note that this uses `*`, which just groups the deps for a
single implementation into one PR. It seems overkill to divvy up each
implementation into granular groups. It makes more sense for this repo
to just have a single dependabot PR for each implementation.